Allow cancellation of index summary redistribution
Patch by Carl Yeksigian; reviewed by marcuse for CASSANDRA-8805
